Changed strings return value from string density to number of strings

+++ include/strings/stringXeon.h
@@ -3,6 +3,6 @@
 
 	#include"scalar/scalarField.h"
 
-	double	stringXeon	(Scalar *axionField, const size_t Lx, const size_t V, const size_t S, FieldPrecision precision, void *string);
-	double	stringCpu	(Scalar *axionField, const size_t Lx, const size_t V, const size_t S, FieldPrecision precision, void *string);
+	size_t	stringXeon	(Scalar *axionField, const size_t Lx, const size_t V, const size_t S, FieldPrecision precision, void *string);
+	size_t	stringCpu	(Scalar *axionField, const size_t Lx, const size_t V, const size_t S, FieldPrecision precision, void *string);
 #endif
